per cycle ac state check
the ac state will be checked every cycle, if it was changed before the tick
(the tick by default is every 10 cycles) then it will force a tick to change
the settings and adapt to the new ac state.

the ac state as always is ignored for desktop type devices.