Improve DESKTOP mode check
A desktop may not even have sensors for the kernel to create the ac_adapter_path in the first place,
if it does then we check for any power supply with a name that starts with B and has type Battery.

This ought to be broad to handle every case i hope.